can upload a bmp as a single frame and display on screen

+++ Loader_esp32bt/buff.h
@@ -14,7 +14,7 @@
 
 /* Size, current position index and byte array of the buffer -----------------*/
 #define Buff__SIZE 4742
-int     Buff__bufInd;
+int16_t     Buff__bufInd;
 char    Buff__bufArr[Buff__SIZE];
 
 /* Reads a byte from the buffer at specified position ------------------------*/
